tweaks

+++ Bug_Popout.cpp
@@ -49,8 +49,8 @@ void Bug_Popout::writeMetaToBug3File(const DataFile &df, const DAQ::BugTask::Blo
 	QTextStream ts(&f);
 	ts << "[ block " << m.blockNum << " ]\n";
 	ts << "framesThisBlock = " << DAQ::BugTask::FramesPerBlock << "\n";
-	ts << "spikeGL_DataFile_ScanCount = " << (df.scanCount()+(fudge/task->numChans())) << "\n";
-	ts << "spikeGL_DataFile_SampleCount = " << (df.sampleCount()+fudge) << "\n";
+	ts << "spikeGL_DataFile_ScanCount = " << (df.scanCount()+u64(fudge/task->numChans())) << "\n";
+	ts << "spikeGL_DataFile_SampleCount = " << (df.sampleCount()+u64(fudge)) << "\n";
 	ts << "spikeGL_ScansInBlock = " << DAQ::BugTask::SpikeGLScansPerBlock << "\n";
 	ts << "boardFrameCounter = ";
 	for (int i = 0; i < DAQ::BugTask::FramesPerBlock; ++i) {
